requirements.txt
msal>=1.24.0
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Dependencies for "Mark as Dealt With" API server (optional)
flask>=2.3.0
flask-cors>=4.0.0